Makefile

+++ download_data.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-DEST=/.
 TMP=$(mktemp -d)
-
-wget http://data.neu.ro/ml-recipe-seismic.zip -O $TMP/ml-recipe-seismic.zip
-unzip $TMP/ml-recipe-seismic.zip -d $DEST
-
+[ ! -f $TMP/ml-recipe-seismic.zip ] && wget http://data.neu.ro/ml-recipe-seismic.zip -O $TMP/ml-recipe-seismic.zip
+unzip $TMP/ml-recipe-seismic.zip -d $DATA_PATH
